Update schema reference tests
The “schema” attribute should only be present in the “parameters” and “responses” definitions.

(Looking at http://swagger.io/v2/schema.json)

Eg. "#/definitions/bodyParameter" schema:

    "bodyParameter": {
      ...
      "properties": {
        ...
        "schema": {
          "$ref": "#/definitions/schema"
        }
      },
      "additionalProperties": false
    },

vs. "#/definitions/schema" schema:

    "definitions": {
      ...
      "additionalProperties": {
        "$ref": "#/definitions/schema"
      },
    },

And "#/definitions/schema":

    "schema": {
      ...
      "properties": {
        "$ref": {
          "type": "string"
        },
        ...
      }
    }

Meaning, a parameter should have a "schema" property that is a schema object (which would likely be a primitive or a $ref), but a definition is a schema object in itself.